requirements.txt
python-dotenv
flask
google-search-results
# requests